requirements.txt
requests~=2.32.4
python-dotenv~=1.1.1

pip~=25.1.1
distro~=1.9.0